docs: service: http: Clean up JavaScript example
* Create docs page for HTTP service JavaScript example
* Fix port assignment for http service
* Add static file serving to HTTP service
* Add serving api.js from HTTP service
* Lint files with js-beautify
* Enforce linting in CI style check
