+++ docs/contributing/style.rst
@@ -4,16 +4,28 @@ Style
 This document talks about code formatting, conventions, documentation, and any
 stylistic choices that we adhere to.
 
-File Formatting
----------------
+Try to run the formatters before every commit. This way, if you push up files
+for review, they are easy to read, even if your pull request isn't yet ready to
+merge.
 
-We run the `black <https://github.com/psf/black>`_ formatter on all files. Try
-to run it before every commit. This way, if you push up files for review, they
-are easy to read, even if your pull request isn't yet ready to merge.
+Python File Formatting
+----------------------
+
+Run the `black <https://github.com/psf/black>`_ formatter on all Python files.
 
 .. code-block:: console
 
     $ black .
+
+JavaScript File Formatting
+--------------------------
+
+Run the `js-beautify <https://github.com/beautify-web/js-beautify>`_ formatter
+on all JavaScript files. Use the following options.
+
+.. code-block:: console
+
+    $ js-beautify -r -n -s 2 file_to_format.js
 
 Imports
 -------
