+++ service/http/MANIFEST.in
@@ -1,2 +1,3 @@
 include README.md
 include LICENSE
+include dffml_service_http/api.js
