+++ service/http/dffml_service_http/api.js
@@ -22,7 +22,7 @@ class DFFMLHTTPAPIObjectContext {
 }
 
 class DFFMLHTTPAPISourceContext extends DFFMLHTTPAPIObjectContext {
-  async records(chunk_size) {
+  async records (chunk_size) {
     // TODO https://www.codementor.io/tiagolopesferreira/asynchronous-iterators-in-javascript-jl1yg8la1
     var response = await this.api.request("/source/" + this.label + "/records/" + chunk_size);
 
@@ -31,7 +31,7 @@ class DFFMLHTTPAPISourceContext extends DFFMLHTTPAPIObjectContext {
     return response.records;
   }
 
-  async update(record) {
+  async update (record) {
     // https://developer.mozilla.org/en-US/docs/Web/API/Fetch_API/Using_Fetch
     var response = await this.api.request("/source/" + this.label + "/update/" + record.key, {
       method: 'POST', // *GET, POST, PUT, DELETE, etc.
@@ -51,7 +51,7 @@ class DFFMLHTTPAPISourceContext extends DFFMLHTTPAPIObjectContext {
 }
 
 class DFFMLHTTPAPIModelContext extends DFFMLHTTPAPIObjectContext {
-  async train(sources) {
+  async train (sources) {
     var source_context_names = [];
     for (var sctx of sources) {
       source_context_names.push(sctx.label);
@@ -75,7 +75,7 @@ class DFFMLHTTPAPIModelContext extends DFFMLHTTPAPIObjectContext {
     return response.records;
   }
 
-  async accuracy(sources) {
+  async accuracy (sources) {
     var source_context_names = [];
     for (var sctx of sources) {
       source_context_names.push(sctx.label);
@@ -99,7 +99,7 @@ class DFFMLHTTPAPIModelContext extends DFFMLHTTPAPIObjectContext {
     return response.records;
   }
 
-  async predict(records) {
+  async predict (records) {
     var response = await this.api.request("/model/" + this.label + "/predict/0", {
       method: 'POST',
       mode: 'cors',
@@ -129,7 +129,7 @@ class DFFMLHTTPAPIObject {
     this.config = {};
   }
 
-  async configure(plugin, label, config) {
+  async configure (plugin, label, config) {
     var response = await this.api.request("/configure/" + this.plugin_type + "/" + plugin + "/" + label, {
       method: 'POST', // *GET, POST, PUT, DELETE, etc.
       mode: 'cors', // no-cors, cors, *same-origin
@@ -152,7 +152,7 @@ class DFFMLHTTPAPIObject {
     return response;
   }
 
-  async context(ctx_label) {
+  async context (ctx_label) {
     var response = await this.api.request("/context/" + this.plugin_type + "/" + this.label + "/" + ctx_label);
 
     await response.json();
@@ -178,7 +178,7 @@ class DFFMLHTTPAPI {
     this.endpoint = endpoint;
   }
 
-  async request(path, options) {
+  async request (path, options) {
     const url = this.endpoint + path;
     const response = await fetch(url, options);
     if (response.status < 400) {
@@ -198,7 +198,7 @@ class DFFMLHTTPAPI {
    * path: Path to place uploaded file at
    * file: String of file contents or input_element.files[0]
    */
-  async upload(path, file) {
+  async upload (path, file) {
     const formData = new FormData();
 
     if (typeof file === "string") {
@@ -225,155 +225,3 @@ class DFFMLHTTPAPI {
     return new DFFMLHTTPAPIModel(this);
   }
 }
-
-var runit = async function() {
-  /*
-   * This is an example of how to use the HTTP api from JavaScript
-   *
-   * The following line creates a new API object where we've substitued port
-   * 9090 for port 8080 (which is where the API is listening).
-   */
-  var api = new DFFMLHTTPAPI(window.location.origin.replace(window.location.port, '8080'));
-
-  console.log("Created api", api);
-
-  // Training data and source
-  var training_source = api.source();
-  console.log("Created training_source", training_source);
-
-  var training_csv_contents = "Years,Expertise,Trust,Salary\n";
-  training_csv_contents += "0,1,0.2,10\n";
-  training_csv_contents += "1,3,0.4,20\n";
-  training_csv_contents += "2,5,0.6,30\n";
-  training_csv_contents += "3,7,0.8,40\n";
-
-  await api.upload("my_training_dataset.csv", training_csv_contents);
-  console.log("Uploaded my_training_dataset.csv");
-
-  await training_source.configure("csv", "my_training_dataset", {
-    "source": {
-      "arg": null,
-      "config": {
-        "filename": {
-          "arg": [
-            "my_training_dataset.csv"
-          ],
-          "config": {}
-        }
-      }
-    }
-  });
-  console.log("Configured training_source", training_source);
-
-  var training_sctx = await training_source.context("my_training_dataset_context");
-  console.log("Created training_sctx", training_sctx);
-
-  // Test data and source
-
-  var test_source = api.source();
-  console.log("Created test_source", test_source);
-
-  var test_csv_contents = "Years,Expertise,Trust,Salary\n";
-  test_csv_contents += "4,9,1.0,50\n";
-  test_csv_contents += "5,11,1.2,60\n";
-
-  await api.upload("my_test_dataset.csv", test_csv_contents);
-  console.log("Uploaded my_test_dataset.csv");
-
-  await test_source.configure("csv", "my_test_dataset", {
-    "source": {
-      "arg": null,
-      "config": {
-        "filename": {
-          "arg": [
-            "my_test_dataset.csv"
-          ],
-          "config": {}
-        }
-      }
-    }
-  });
-  console.log("Configured test_source", test_source);
-
-  var test_sctx = await test_source.context("my_test_dataset_context");
-  console.log("Created test_sctx", test_sctx);
-
-  // Create an array of all the records for fun
-  var records = await training_sctx.records(100);
-  console.log("Training records", records);
-
-  var records_array = [];
-  for (var key of Object.keys(records)) {
-    records_array.push(records[key]);
-  }
-  console.log("Array of training records", records_array);
-
-  // Create a model
-  var model = api.model();
-  console.log("Created model", model);
-
-  await model.configure("scikitlr", "mymodel", {
-    "model": {
-      "arg": null,
-      "config": {
-        "predict": {
-          "arg": [
-            "Salary"
-          ],
-          "config": {}
-        },
-        "features": {
-          "arg": [
-            {
-              "name": "Years",
-              "dtype": "int",
-              "length": 1
-            },
-            {
-              "name": "Expertise",
-              "dtype": "int",
-              "length": 1
-            },
-            {
-              "name": "Trust",
-              "dtype": "float",
-              "length": 1
-            }
-          ],
-          "config": {}
-        }
-      }
-    }
-  });
-
-  console.log("Configured model", model);
-
-  var mctx = await model.context("mymodel_context");
-  console.log("Created model context", mctx);
-
-  await mctx.train([training_sctx]);
-  console.log("Trained model context", mctx);
-
-  var accuracy = await mctx.accuracy([test_sctx]);
-  console.log("Model context accuracy", accuracy);
-
-  var prediction = await mctx.predict({
-    "mish_the_smish": {
-      "features": {
-        "Years": 6,
-        "Expertise": 13,
-        "Trust": 1.4
-      }
-    }
-  });
-  console.log("Model context predict", prediction);
-
-  if (prediction.mish_the_smish.prediction.value !== 70) {
-    console.error(prediction)
-    throw new Error("prediction.mish_the_smish.prediction.value was not 70!");
-  }
-
-  console.log("Success!");
-}
-
-runit();
