+++ service/http/docs/index.rst
@@ -46,3 +46,4 @@ You should instead see the :doc:`security` docs!
 
     api
     security
+    javascript
