+++ service/http/examples/javascript/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="UTF-8">
+    <script src="api.js" ></script>
+    <script src="index.js" ></script>
+  </head>
+  <body>
+    <h1>Open the console to see logs</h1>
+  </body>
+</html>
