+++ service/http/examples/javascript/setup.sh
