--- service/http/examples/web/index.html
@@ -1,6 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <body>
-    <script src="api.js" ></script>
-  </body>
-</html>
